requirements.txt
PySide6
opencv-python
onnxruntime
pyyaml